Ai_Quiz_Generator/requirements.txt
flask
streamlit
pdfplumber
transformers
torch
sentencepiece
fpdf2
requests
python-dotenv
accelerate